+++ ui/app/components/transaction-list.js
@@ -1,14 +1,21 @@
+const Component = require('react').Component
 const h = require('react-hyperscript')
-const vreme = new (require('vreme'))
-const formatBalance = require('../util').formatBalance
-const addressSummary = require('../util').addressSummary
-const explorerLink = require('../../lib/explorer-link')
-const Panel = require('./panel')
-const Identicon = require('./identicon')
-const EtherBalance = require('./eth-balance')
+const inherits = require('util').inherits
+
+const TransactionListItem = require('./transaction-list-item')
+
+module.exports = TransactionList
 
 
-module.exports = function(transactions, network) {
+inherits(TransactionList, Component)
+function TransactionList() {
+  Component.call(this)
+}
+
+TransactionList.prototype.render = function() {
+  const { txsToRender, network, unconfTxs, unconfMsgs } = this.props
+  const transactions = txsToRender
+
   return (
 
     h('section.transaction-list', [
@@ -42,118 +49,19 @@ module.exports = function(transactions, network) {
       }, (
 
         transactions.length ?
-          transactions.map(renderTransaction)
+          transactions.map((transaction, i) => {
+            return h(TransactionListItem, {
+              transaction, i
+            })
+          })
         :
           [h('.flex-center', {
             style: {
               height: '100%',
             },
           }, 'No transaction history...')]
-
       ))
-
     ])
-
   )
-
-
-  function renderTransaction(transaction, i){
-
-    var txParams = transaction.txParams
-    var date = formatDate(transaction.time)
-
-    return (
-
-      h(`.transaction-list-item.flex-row.flex-space-between${transaction.hash ? '.pointer' : ''}`, {
-        key: `tx-${transaction.id + i}`,
-        onClick: (event) => {
-          if (!transaction.hash) return
-          var url = explorerLink(transaction.hash, parseInt(network))
-          chrome.tabs.create({ url })
-        },
-        style: {
-          padding: '20px 0',
-        },
-      }, [
-
-        // large identicon
-        h('.identicon-wrapper.flex-column.flex-center.select-none', [
-          identicon(txParams, transaction),
-        ]),
-
-        h('.flex-column', [
-
-          h('div', date),
-
-          recipientField(txParams, transaction),
-
-        ]),
-
-        h(EtherBalance, {
-          value: txParams.value,
-        }),
-      ])
-
-    )
-  }
 }
 
-function recipientField(txParams, transaction) {
-  if (txParams.to) {
-    return h('div', {
-      style: {
-        fontSize: 'small',
-        color: '#ABA9AA',
-      },
-    }, [
-      addressSummary(txParams.to),
-      failIfFailed(transaction),
-    ])
-
-  } else {
-
-    return h('div', {
-      style: {
-        fontSize: 'small',
-        color: '#ABA9AA',
-      },
-    },[
-      'Contract Published',
-      failIfFailed(transaction),
-    ])
-
-  }
-}
-
-function formatDate(date){
-  return vreme.format(new Date(date), 'March 16 2014 14:30')
-}
-
-function identicon(txParams, transaction) {
-  if (transaction.status === 'rejected') {
-    return h('i.fa.fa-exclamation-triangle.fa-lg.error', {
-      style: {
-        width: '24px',
-      }
-    })
-  }
-
-  if (txParams.to) {
-    return h(Identicon, {
-      diameter: 24,
-      address: txParams.to || transaction.hash,
-    })
-  } else {
-    return h('i.fa.fa-file-text-o.fa-lg', {
-      style: {
-        width: '24px',
-      }
-    })
-  }
-}
-
-function failIfFailed(transaction) {
-  if (transaction.status === 'rejected') {
-    return h('span.error', ' (Failed)')
-  }
-}